requirements.txt
bleak
Pillow
python-dotenv